evfleet-ws/requirements.txt
aws-cdk-lib
aws-cdk.aws-lambda-python-alpha
constructs>=10.0.0,<11.0.0
pyyaml==6.0
